wip: refactor request validator
