+++ src/framework/http/validator.ts
@@ -1,4 +1,6 @@
+import { type RequestValidator } from '@/interface/http/request-validator.type';
 import { builtInConstraints, createValidationMiddleware, createValidator } from '@koala-ts/framework/validator';
 
 const validate = createValidator({ constraints: builtInConstraints });
-export const validateRequest = createValidationMiddleware({ validate });
+
+export const validateRequest: RequestValidator = createValidationMiddleware({ validate });
