+++ src/interface/http/request-validator.type.ts
@@ -1,0 +1,4 @@
+import { type HttpMiddleware } from '@koala-ts/framework';
+import { type ValidationSchema } from '@koala-ts/framework/validator';
+
+export type RequestValidator = (schema: ValidationSchema) => HttpMiddleware;
