+++ src/kairos/workspace/workspace-composition.ts
@@ -1,5 +1,5 @@
+import { validateRequest } from '@/framework/http/validator';
 import { workspacesCollection } from '@/framework/mongodb/collection/workspaces-collection';
-import { validateRequest } from '@/kairos/shared/http/request-validation';
 import { createCreateWorkspace } from '@/kairos/workspace/application/create-workspace';
 import { type SaveWorkspace } from '@/kairos/workspace/domain/workspace-repository';
 import { createSaveWorkspace } from '@/kairos/workspace/infrastructure/repository/save-workspace';
